the pdpt index reaches above the 32-bit boundary too
The pml4 index was widened with the tlb, but the pdpt index below it was still taken from the
address truncated to 32 bits. It is bits 38:30, so truncation keeps only the bottom two and drops
the rest - which sends the walk to the wrong entry and faults on a page that is mapped.

The fixture did not catch this, because the address it used had a pdpt index that survived
truncation. That made it a test of the pml4 index alone. It now uses 517 GiB, whose pdpt index is
5 and truncates to 1, and fails with exactly the page fault windows was taking when the fix is
backed out.

+++ src/rust/cpu/cpu.rs
@@ -2455,7 +2455,8 @@ pub unsafe fn do_page_walk(
             allow_user &= pml4_entry as i32 & PAGE_TABLE_USER_MASK != 0;
 
             let pdpt_entry_addr =
-                (pml4_entry as u32 & 0xFFFFF000) + ((((addr as u32) >> 30) & 0x1FF) << 3);
+                (pml4_entry as u32 & 0xFFFFF000)
+                    + (((((addr as u64) >> 30) & 0x1FF) as u32) << 3);
             let pdpt_entry = memory::read64s(pdpt_entry_addr);
 
             if pdpt_entry as i32 & PAGE_TABLE_PRESENT_MASK == 0 {
